+++ src/osha/oira/client/__init__.py
@@ -1,0 +1,2 @@
+# Make sure sqlalchemy subscribers are initialized
+from .subscribers import update_user_languages_subscriber  # noqa: F401